Move mapSymbols sync-schema to separate file
